+++ lib/rest/chat.dart
@@ -20,4 +20,24 @@ abstract class _ClientChatMixin implements _ClientWrapper {
     }).catchError((error) => completer.completeError(error));
     return completer.future;
   }
+
+  Future<void> reactMessage(String messageId, String emoji,
+      [bool shouldReact]) {
+    Completer<void> completer = Completer();
+    final body = json.encode(<String, dynamic>{
+      'messageId': messageId,
+      'emoji': emoji,
+      'shouldReact': shouldReact,
+    });
+    final headers = <String, String>{
+      'X-User-Id': _auth._id,
+      'X-Auth-Token': _auth._token,
+      'Content-Type': 'application/json',
+    };
+    http
+        .post('${_getUrl()}/chat.react', headers: headers, body: body)
+        .then((response) => completer.complete(null))
+        .catchError((error) => completer.completeError(error));
+    return completer.future;
+  }
 }