+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.31 (2018-10-27)
+
+### Feature:
+
+* Add react message.
+* Add Message's *reactions* fields
+
+
 ## 0.0.30 (2018-10-24)
 
 ### Feature: